develop코드 merge

See merge request s07-bigdata-recom-sub2/S07P22A107!217

Former-commit-id: 2611c4e3378a03c047ed73633586f96b883e6eff

+++ BE/src/main/java/com/team7/chaekin/domain/meeting/service/MeetingService.java
@@ -9,6 +9,7 @@ import com.team7.chaekin.domain.member.entity.Member;
 import com.team7.chaekin.domain.member.repository.MemberRepository;
 import com.team7.chaekin.domain.participant.entity.Participant;
 import com.team7.chaekin.domain.participant.repository.ParticipantRepository;
+import com.team7.chaekin.global.error.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,6 +18,8 @@ import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import static com.team7.chaekin.global.error.errorcode.DomainErrorCode.*;
 
 @RequiredArgsConstructor
 @Service
@@ -29,7 +32,7 @@ public class MeetingService {
 
     @Transactional
     public MeetingListResponse getMeetings(MeetingListRequest meetingListRequest, Pageable pageable) {
-        Page<Meeting> pageList = meetingRepository.findByTitleContaining(meetingListRequest.getKeyword(), pageable);
+        Page<Meeting> pageList = meetingRepository.searchMeetingList(meetingListRequest.getKeyword(), pageable);
 
         List<MeetingListDto> meetings = pageList.toList().stream()
                 .map(Meeting::toListDto)
@@ -39,21 +42,35 @@ public class MeetingService {
     }
 
     @Transactional
-    public MeetingDetailResponse getMeetingDetail(long meetingId) {
-        return getMeeting(meetingId).toDetailDto();
+    public MeetingDetailResponse getMeetingDetail(long meetingId, long memberId) {
+        Meeting meeting = getMeeting(meetingId);
+        MeetingDetailResponse meetingDetailResponse = meeting.toDetailDto();
+
+        Member meetingLeader = meeting.getMeetingLeader();
+        meetingDetailResponse.setIsMine(meetingLeader.getId().equals(memberId));
+        return meetingDetailResponse;
+    }
+
+    @Transactional
+    public MeetingMyResponse getMyMeetings(long memberId) {
+        List<Participant> participants = participantRepository.findByMemberId(memberId);
+        List<MeetingListDto> dtoList = participants.stream()
+                .map(participant -> participant.getMeeting())
+                .map(Meeting::toListDto)
+                .collect(Collectors.toList());
+        return new MeetingMyResponse(dtoList);
     }
 
     @Transactional
     public long createMeeting(long memberId, MeetingCreateRequest meetingCreateRequest) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new RuntimeException("해당 회원이 존재하지 않습니다."));
-        Book book = bookRepository.findById(meetingCreateRequest.getBookId())
-                .orElseThrow(() -> new RuntimeException("해당 책이 존재하지 않습니다."));
+        Member member = getMember(memberId);
+        Book book = getBook(meetingCreateRequest.getBookId());
 
         Meeting meeting = meetingRepository.save(Meeting.builder()
                 .book(book)
                 .title(meetingCreateRequest.getTitle())
                 .description(meetingCreateRequest.getDescription())
+                .meetingStatus(meetingCreateRequest.getMeetingStatus())
                 .capacity(meetingCreateRequest.getMaxCapacity()).build());
 
         participantRepository.save(Participant.makeParticipant(meeting, member, true));
@@ -61,23 +78,41 @@ public class MeetingService {
     }
 
     @Transactional
-    public void updateMeeting(long meetingId, MeetingUpdateRequest meetingUpdateRequest) {
+    public void updateMeeting(long meetingId, long memberId, MeetingUpdateRequest meetingUpdateRequest) {
         Meeting meeting = getMeeting(meetingId);
-        Book book = bookRepository.findById(meetingUpdateRequest.getBookId())
-                .orElseThrow(() -> new RuntimeException("해당 책이 존재하지 않습니다."));
+        Book book = getBook(meetingUpdateRequest.getBookId());
 
+        if (meetingUpdateRequest.getMaxCapacity() < meeting.getCurrentParticipants()) {
+            throw new CustomException(IMPOSSIBLE_CAPACITY_LESS_THAN_CURRENT_MEMBERS);
+        }
+        if (!meeting.getMeetingLeader().getId().equals(memberId)) {
+            throw new CustomException(DO_NOT_HAVE_AUTHORIZATION);
+        }
         meeting.update(book, meetingUpdateRequest);
     }
 
     @Transactional
-    public void deleteMeeting(long meetingId) {
+    public void deleteMeeting(long memberId, long meetingId) {
         Meeting meeting = getMeeting(meetingId);
+        if (!meeting.getMeetingLeader().getId().equals(memberId)) {
+            throw new CustomException(ONLY_LEADER_CAN_DELETE_MEETING);
+        }
         meeting.delete();
     }
 
-    private Meeting getMeeting(long meetingId) {
-        return meetingRepository.findById(meetingId)
-                .orElseThrow(() -> new RuntimeException("해당 모임이 존재하지 않습니다."));
+    private Book getBook(long bookId) {
+        return bookRepository.findById(bookId)
+                .orElseThrow(() -> new CustomException(BOOK_IS_NOT_EXIST));
     }
 
+    private Meeting getMeeting(long meetingId) {
+        return meetingRepository.findByIdAndIsRemovedIsFalse(meetingId)
+                .orElseThrow(() -> new CustomException(MEETING_IS_NOT_EXIST));
+    }
+
+    private Member getMember(long memberId) {
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new CustomException(MEMBER_IS_NOT_EXIST));
+        return member;
+    }
 }
